Changed the approach for pull/push authorization.
Now we capture the git-receive-pack requests and rewrite them with a /push prefix and force a pass through.

This now allows a regular Location type directive to be setup for the push location and thus can be used for
authorizing the push and also enforce the network restrictions.

For the time being we use 2 ScriptAliasMatch entries to pass the repository URL through to the git-http-backend.